EWallet
"Creates list of incomes, expenses, and months. It also has the options to print income report by getting monthly information, print expense report for the specific month, print savings for the month. It displays each of the 3 and outputs the grand total by savings - expenses for the month = grand total. Final feature was the overall report which prints all 3.